Created utils for training function
